==> 2 - train_test_val_split.ipynb <==
# Train Test Validation Split

import pickle
import pandas as pd
from sklearn.model_selection import train_test_split 

df = pd.read_pickle("dataframes/pre-processed.pkl")

df.describe()

# Shuffle and split dataset randomly into 80% training and 20% test, then split training into 80% training and 20% validation
train, test = train_test_split(df, test_size=0.2, random_state=42)
train, val = train_test_split(train, test_size=0.2, random_state=42)

# Validate that the split is correct
print("Train size: ", len(train))
print("Validation size: ", len(val))
print("Test size: ", len(test))

# Validate that the split is a representative sample of the original dataset
print("Train mean: ", train["adjusted_price"].mean())
print("Validation mean: ", val["adjusted_price"].mean())
print("Test mean: ", test["adjusted_price"].mean())

# Split into X and y
X_train = train.drop(columns=["adjusted_price"])
y_train = train["adjusted_price"]

X_val = val.drop(columns=["adjusted_price"])
y_val = val["adjusted_price"]

X_test = test.drop(columns=["adjusted_price"])
y_test = test["adjusted_price"]

X_train.to_pickle("dataframes/x-train.pkl")
y_train.to_pickle("dataframes/y-train.pkl")

X_val.to_pickle("dataframes/x-val.pkl")
y_val.to_pickle("dataframes/y-val.pkl")

X_test.to_pickle("dataframes/x-test.pkl")
y_test.to_pickle("dataframes/y-test.pkl")
