==> 3 - model_selection.ipynb <==
# Training Models

from sklearn.metrics import mean_squared_error, mean_absolute_error
from sklearn.metrics import mean_absolute_percentage_error
import pandas as pd
import pickle
import numpy as np

# Load y_train, y_test, X_train, X_test

y_train = pd.read_pickle('dataframes/y-train.pkl')
y_val = pd.read_pickle('dataframes/y-val.pkl')
X_train = pd.read_pickle('dataframes/x-train.pkl')
X_val = pd.read_pickle('dataframes/x-val.pkl')

# Get MSE, MAE and MPE for model where we predict the mean of the training set
mean = y_train.mean()
y_pred = np.full(len(y_val), mean)

print("MSE: ", mean_squared_error(y_val, y_pred))
print("MAE: ", mean_absolute_error(y_val, y_pred))
print("MPE: ", mean_absolute_percentage_error(y_val, y_pred))

# # Calculate f_regression scores for each feature
# from sklearn.feature_selection import f_regression

# # Sort feature names by f_regression scores
# feature_names = X_train.columns
# scores = f_regression(X_train, y_train)[0]
# feature_scores = pd.DataFrame({"feature": feature_names, "score": scores})
# feature_scores.sort_values(by="score", ascending=False)


# Train Linear Regression and evaluate on validation set
from sklearn.linear_model import LinearRegression
lin_reg = LinearRegression()
lin_reg.fit(X_train, y_train)

# Print MSE and MAE
y_val_pred = lin_reg.predict(X_val)
print("MSE for linear regression model 1 =>", mean_squared_error(y_val, y_val_pred))
print("MAE for linear regression model 1 =>", mean_absolute_error(y_val, y_val_pred))
print("MAPE for linear regression model 1 =>", mean_absolute_percentage_error(y_val, y_val_pred))


# Train Ridge Regression and evalue on validation set
from sklearn.linear_model import Ridge

rr = Ridge(alpha=1).fit(X_train, y_train) 

y_val_pred = rr.predict(X_val)
print("MSE for ridge regression alpha 1 =>", mean_squared_error(y_val, y_val_pred))
print("MAE for ridge regression alpha 1 =>", mean_absolute_error(y_val, y_val_pred))

rr = Ridge(alpha=10).fit(X_train, y_train) 

y_val_pred = rr.predict(X_val)
print("MSE for ridge regression alpha 10 =>", mean_squared_error(y_val, y_val_pred))
print("MAE for ridge regression alpha 10 =>", mean_absolute_error(y_val, y_val_pred))

rr = Ridge(alpha=100).fit(X_train, y_train) 

y_val_pred = rr.predict(X_val)
print("MSE for ridge regression alpha 100 =>", mean_squared_error(y_val, y_val_pred))
print("MAE for ridge regression alpha 100 =>", mean_absolute_error(y_val, y_val_pred))

rr = Ridge(alpha=1000).fit(X_train, y_train) 

y_val_pred = rr.predict(X_val)
print("MSE for ridge regression alpha 1000 =>", mean_squared_error(y_val, y_val_pred))
print("MAE for ridge regression alpha 1000 =>", mean_absolute_error(y_val, y_val_pred))

rr = Ridge(alpha=10000).fit(X_train, y_train) 

y_val_pred = rr.predict(X_val)
print("MSE for ridge regression alpha 10000 =>", mean_squared_error(y_val, y_val_pred))
print("MAE for ridge regression alpha 10000 =>", mean_absolute_error(y_val, y_val_pred))

rr = Ridge(alpha=100000).fit(X_train, y_train) 

y_val_pred = rr.predict(X_val)
print("MSE for ridge regression alpha 10000 =>", mean_squared_error(y_val, y_val_pred))
print("MAE for ridge regression alpha 10000 =>", mean_absolute_error(y_val, y_val_pred))


# # Train support vector regression and evaluate on validation set
from sklearn.svm import SVR

svr = SVR(kernel="linear", C=1, epsilon=0.1)
svr.fit(X_train, y_train)

y_val_pred = svr.predict(X_val)
print("MSE for SVR =>", mean_squared_error(y_val, y_val_pred))
print("MAE for SVR =>", mean_absolute_error(y_val, y_val_pred))

# Train Decison Tree Regressor and evaluate on validation set
from sklearn.tree import DecisionTreeRegressor

tree_reg = DecisionTreeRegressor(random_state=42).fit(X_train, y_train)

y_val_pred = tree_reg.predict(X_val)
print("MSE for decision tree regression model 1 =>", mean_squared_error(y_val, y_val_pred))
print("MAE for decision tree regression model 1 =>", mean_absolute_error(y_val, y_val_pred))
print("MAPE for decision tree regression model 1 =>", mean_absolute_percentage_error(y_val, y_val_pred))

print(tree_reg.get_depth())

# train using k-nearest neighbors
# from sklearn.neighbors import KNeighborsRegressor

# knn_reg = KNeighborsRegressor(weights="distance", n_jobs=8).fit(X_train, y_train)

# y_val_pred = knn_reg.predict(X_val)
# print("MSE for k-nearst neighbors regression model 1 =>", mean_squared_error(y_val, y_val_pred))
# print("MAE for k-nearst neighbors regression model 1 =>", mean_absolute_error(y_val, y_val_pred))
# print("MAPE for k-nearst neighbors regression model 1 =>", mean_absolute_percentage_error(y_val, y_val_pred))

# # Train Linear support vector regressor and evaluate on validation set
# from sklearn.svm import LinearSVR

# svm_reg = LinearSVR(random_state=42).fit(X_train, y_train)

# y_val_pred = svm_reg.predict(X_val)
# print("MSE for linear support vector regression model 1 =>", mean_squared_error(y_val, y_val_pred))
# print("MAE for linear support vector regression model 1 =>", mean_absolute_error(y_val, y_val_pred))

# Train Random Forest Regressor and evaluate on validation set
from sklearn.ensemble import RandomForestRegressor

forest_reg = RandomForestRegressor(n_estimators=100, random_state=42, n_jobs=8).fit(X_train, y_train)

y_val_pred = forest_reg.predict(X_val)
print("MSE for random forest regression model 1 =>", mean_squared_error(y_val, y_val_pred))
print("MAE for random forest regression model 1 =>", mean_absolute_error(y_val, y_val_pred))
print("MAPE for random forest regression model 1 =>", mean_absolute_percentage_error(y_val, y_val_pred))

max_depth = 0
for dt_reg in forest_reg.estimators_:
	if dt_reg.get_depth() > max_depth:
		max_depth = dt_reg.get_depth()

print(max_depth)
print(len(forest_reg.estimators_))
