==> 4 - random_forest_hyperparameter_tuning.ipynb <==
# Random Forest Hyperparameter Tuning

# Import Libraries
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_squared_error, mean_absolute_error
import math

# Import train and validation data
y_train = pd.read_pickle('dataframes/y-train.pkl')
y_val = pd.read_pickle('dataframes/y-val.pkl')
X_train = pd.read_pickle('dataframes/x-train.pkl')
X_val = pd.read_pickle('dataframes/x-val.pkl')

# Number of Estimators Hyperparameter Tuning
errors = []
growing_rf = RandomForestRegressor(n_estimators=1, n_jobs=6,  
                                    warm_start=True, random_state=42)
for i in [1, 5, 10, 20, 25, 30, 40, 50, 60, 70, 80, 100, 150, 200, 250]:
    growing_rf.n_estimators=i
    growing_rf.fit(X_train, y_train)
    predicted = growing_rf.predict(X_val)
    errors.append(mean_squared_error(y_val, predicted))
    print('Iteration: ', i, 'RMSE Loss: ', math.sqrt(errors[-1]), ' MAE: ', mean_absolute_error(y_val, predicted))

# Max Depth Hyperparameter Tuning
from sklearn.ensemble import RandomForestRegresso

errors = []
for i in [1, 5, 10, 15, 20, 25,30, 40, 60, 70]:
    growing_rf = RandomForestRegressor(n_estimators=75, n_jobs=6, random_state=42, max_depth=i)
    growing_rf.fit(X_train, y_train)
    predicted = growing_rf.predict(X_val)
    errors.append(mean_squared_error(y_val, predicted))
    print('Iteration: ', i, 'RMSE Loss: ', math.sqrt(errors[-1]), ' MAE: ', mean_absolute_error(y_val, predicted))
